i|Remove fieldAddonLeft and fieldAddonRight for performance reasons
Rendering massive schemas/forms this simple change results in 400+ calls
to ngIfWatchAction instead of 800+ calls. In our case, this reduces
time spent within ngIfWatchAction from ~600ms to ~100ms.
Should fieldAddonLeft or fieldAddonRight ever be needed it would be
better to implement it as a builder, as this would greatly reduce the
runtime cost of angularJS directives. There is no reason for bootstrap
input-groups to be dependent on evaluation by ng-if, as we could do this
during the build process of a single form field.