b|Fix jump-to-buttons
In their current state the jump-to-buttons didn't work as expected. This
problem stems from the fact that offsetTop is calculated only in regard
to the offsetParent of an element. As we need the offset in regard to
the complete document body, we switch to the offset() function of
jquery, which calculates exactly that.